requirements.txt
python-dotenv==1.0.0
pexels-api==1.0.1
moviepy==1.0.3
openai==1.3.0
python-ffmpeg==2.0.8
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0 